fix: user_id 타입 변경

+++ backend/src/main/java/com/a407/back/model/service/basicService.java
@@ -13,7 +13,7 @@ import java.util.Optional;
 public class basicService {
     private final basicRepository basicRepository;
     @Transactional
-    public Long save(User user) {
+    public int save(User user) {
         return basicRepository.save(user).getId();
     }
 }
